fix: renamed .t.ts do .d.ts
