+++ src/index.d.ts
@@ -1,0 +1,1 @@
+export { UserConfig } from './types/config';
