--- src/index.ts
@@ -1,1 +1,0 @@
-export { UserConfig } from './types/config.t';
