+++ src/types/config.d.ts
